(maint) Update makefile to reset go.mod changes in implements
The goal of implements written in go is to use the most recent regulator
when building. In the case of publication the most recent regulator
changes, so instead of figuring out how to update the version number in
each impl we just never commit the required go packages for impls and
let 'go mod tidy' take care of getting the latest regulator.

Then whenever we build implements we reset any changes 'go mod tidy'
made so that the next time tidy runs it gets the newest version.